requirements.txt
fake-useragent
requests
progressbar2


